Add create user cart endpoint
Adds user cart documents and DTO to the service, user carts will store books before they move on the order service. Adds and endpoint to create the cart

Before creating the cart must validate that the amounts are within the stored stock amount of books and that the amount of books to add is greater than 0. Add validation and test cases for both scenarios